--- a/src/phyLiNCoptimization.jl
+++ b/src/phyLiNCoptimization.jl
@@ -52,7 +52,7 @@ See constructor below, from a `StatisticalSubstitutionModel`.
 struct CacheLengthLiNC
     """forward likelihood of descendants of focus edge: nstates x nsites x nrates x ntrees"""
     flike::Array{Float64,4}
-    """likelihood of non-descendants of focus edge: direct * backwards.
+    """likelihood of non-descendants of focus edge: direct * backwards * P(rate & tree).
        size: nstates x nsites x nrates x ntrees"""
     dblike::Array{Float64,4}
     """whether displayed trees have the focus edge (true) or not (false),
@@ -712,6 +712,7 @@ function nni_LiNC!(obj::SSM, no3cycle::Bool, nohybridladder::Bool,
         updateSSM!(obj, true; constraints=constraints)
         optimizelocalgammas_LiNC!(obj, e1, ftolAbs, γcache)
         # don't delete a hybrid edge with γ=0: to be able to undo the NNI
+        # but: there's no point in optimizing the length of such an edge
         optimizelocalBL_LiNC!(obj, e1, lcache)
         if obj.loglik < currLik
             nni!(undoinfo...) # undo move
@@ -890,7 +891,7 @@ function deletehybridedgeLiNC!(obj::SSM, currLik::Float64,
     majhyb = getMajorParentEdge(hybridnode)
     len0 = majhyb.length # to restore later if deletion rejected
     update_logtrans(obj) # fixit: is that really needed?
-    optimizelength_LiNC!(obj, majhyb, lcache, Q(obj.model), exp.(obj.priorltw))
+    optimizelength_LiNC!(obj, majhyb, lcache, Q(obj.model))
     # don't optimize gammas: because we want to constrain one of them to 0.0
     if obj.loglik - currLik > likAbsDelHybLiNC # -0.1: loglik can decrease for parsimony
         deletehybridedge!(obj.net, minorhybridedge, false,true,false,false,false) # nofuse,unroot,multgammas,simplify
@@ -1198,6 +1199,8 @@ function updatecache_edge!(lcache::CacheLengthLiNC, obj::SSM, focusedge)
         @debug "edge $(focusedge.number) does not affect the likelihood: skip optimization"
         return missing
     end
+    lrw  = obj.ratemodel.lograteweight
+    ltw  = obj.priorltw
     ftmp = obj.forwardlik
     btmp = obj.backwardlik
     dtmp = obj.directlik
@@ -1205,11 +1208,13 @@ function updatecache_edge!(lcache::CacheLengthLiNC, obj::SSM, focusedge)
     fill!(flike, -Inf); fill!(dblike, -Inf)
     for it in 1:nt
         if hase[it] # we do need flike and dblike
-        for ir in 1:nr for is in 1:ns
-            clik[is,ir,it] = discrete_corelikelihood_trait!(obj, it,is,ir)
+        for ir in 1:nr
+          ltrw = ltw[it] + lrw[ir]
+          for is in 1:ns
+            clik[is,ir,it] = discrete_corelikelihood_trait!(obj, it,is,ir) # ! +ltrw not needed
             discrete_backwardlikelihood_trait!(obj, it,ir)
             flike[:,is,ir,it]  .= ftmp[:,unum]
-            dblike[:,is,ir,it] .= btmp[:,vnum]
+            dblike[:,is,ir,it] .= btmp[:,vnum] .+ ltrw
             for isis in 1:nsis
             @inbounds if hassis[isis,it]
                 dblike[:,is,ir,it] .+= dtmp[:,snum[isis]] # sum on log scale
@@ -1217,8 +1222,10 @@ function updatecache_edge!(lcache::CacheLengthLiNC, obj::SSM, focusedge)
             end
         end; end
         else # tree does't have the focus edge: use clik only
-        for ir in 1:nr for is in 1:ns
-            clik[is,ir,it] = discrete_corelikelihood_trait!(obj, it,is,ir)
+        for ir in 1:nr
+          ltrw = ltw[it] + lrw[ir]
+          for is in 1:ns
+            clik[is,ir,it] = discrete_corelikelihood_trait!(obj, it,is,ir) + ltrw
         end; end
         end
     end
@@ -1255,10 +1262,9 @@ function optimizealllengths_LiNC!(obj::SSM, lcache::CacheLengthLiNC)
     end
     update_logtrans(obj)
     qmat  = Q(obj.model)
-    ltw = exp.(obj.priorltw)
     Random.shuffle!(edges)
     for e in edges
-        optimizelength_LiNC!(obj, e, lcache, qmat, ltw)
+        optimizelength_LiNC!(obj, e, lcache, qmat)
     end
     return edges # excludes constrained edges
 end
@@ -1309,8 +1315,8 @@ julia> writeTopology(obj.net; round=true)
 function optimizelocalBL_LiNC!(obj::SSM, focusedge::Edge, lcache::CacheLengthLiNC)
     neighboredges = adjacentedges(focusedge)
     # remove from neighboredges (shallow copy!)
-    #  - constrained edges: to keep network unzipped, add
-    #  - and any hybrid edge with γ=0
+    #  - constrained edges: to keep network unzipped, and
+    #  - any hybrid edge with γ=0
     if !isempty(obj.net.hybrid)
         @inbounds for i in length(neighboredges):-1:1
             e = neighboredges[i]
@@ -1324,16 +1330,14 @@ function optimizelocalBL_LiNC!(obj::SSM, focusedge::Edge, lcache::CacheLengthLiN
     # assume: correct preordered displayed trees & tree weights
     update_logtrans(obj) # fixit: waste of time? avoid in some cases?
     qmat  = Q(obj.model)
-    ltw = exp.(obj.priorltw)
     for e in neighboredges
-        optimizelength_LiNC!(obj, e, lcache, qmat, ltw)
+        optimizelength_LiNC!(obj, e, lcache, qmat)
     end
     return neighboredges # excludes constrained edges
 end
 
 """
-    optimizelength_LiNC!(obj::SSM, edge::Edge, cache::CacheLengthLiNC,
-                         Qmatrix, treeweights)
+    optimizelength_LiNC!(obj::SSM, edge::Edge, cache::CacheLengthLiNC, Qmatrix)
 
 Optimize the length of a single `edge` using a gradient method, if
 `edge` is not below a reticulation (the unzipped canonical version should
@@ -1343,14 +1347,16 @@ Output: nothing.
 Warning: displayed trees are assumed up-to-date, with nodes preordered
 """
 function optimizelength_LiNC!(obj::SSM, focusedge::Edge,
-                          lcache::CacheLengthLiNC, qmat, ltw)
+                              lcache::CacheLengthLiNC, qmat)
     if getParent(focusedge).hybrid # keep the reticulation unzipped
         return nothing # the length of focus edge should be 0. stay as is.
     end
     cfg = updatecache_edge!(lcache, obj, focusedge)
     ismissing(cfg) && return nothing
-    flike  = lcache.flike
-    dblike = lcache.dblike
+    flike  = lcache.flike  # P(descendants of e | state at child, rate, tree)
+    dblike = lcache.dblike # P(non-descendants & state at parent, rate, tree)
+    # if γ=0 then dblike starts as all -Inf because P(tree) = 0 -> problem
+    # this problem is caught earlier: cfg would have been missing.
     hase   = lcache.hase
     Prt    = lcache.Prt
     rQP    = lcache.rQP
@@ -1359,9 +1365,9 @@ function optimizelength_LiNC!(obj::SSM, focusedge::Edge,
     k, ns, nr, nt = size(flike)
     nt = length(tree) # could be less than dimension in lcache
     rates = obj.ratemodel.ratemultiplier
-    ulik  = obj._sitecache   # unconditional likelihood
-    clik  = obj._loglikcache # conditional likelihood
-    adjustment = obj.totalsiteweight * (cfg - log(nr))
+    ulik  = obj._sitecache   # will hold P(site): depends on length of e
+    clik  = obj._loglikcache # P(site & rate, tree) using old length of e
+    adjustment = obj.totalsiteweight * cfg
 
     # use site weights, if any
     wsum = (isnothing(obj.siteweight) ? sum : x -> sum(x .* obj.siteweight))
@@ -1378,26 +1384,21 @@ function optimizelength_LiNC!(obj::SSM, focusedge::Edge,
         for it in 1:nt
         if hase[it] # tree has edge: contributes to gradient
             for is in 1:ns
-                u=0.0; g=0.0
                 for ir in 1:nr
                     # dot(x,A,y) requires Julia 1.4, and fails with Optim v0.19.3
                     # u += dot(dblike[:,is,ir,it], Prt[ir], flike[:,is,ir,it])
                     # g += dot(dblike[:,is,ir,it], rQP[ir], flike[:,is,ir,it])
                     @inbounds for i in 1:k for j in 1:k
-                        u += dblike[i,is,ir,it] * Prt[ir][i,j] * flike[j,is,ir,it]
-                        g += dblike[i,is,ir,it] * rQP[ir][i,j] * flike[j,is,ir,it]
+                        ulik[is] += dblike[i,is,ir,it] * Prt[ir][i,j] * flike[j,is,ir,it]
+                        glik[is] += dblike[i,is,ir,it] * rQP[ir][i,j] * flike[j,is,ir,it]
                     end; end
                 end
-                ulik[is] += u * ltw[it]
-                glik[is] += g * ltw[it]
             end
         else # tree doesn't have focus edge: use clik only, gradient=0
             for is in 1:ns
-                u=0.0
                 for ir in 1:nr
-                    u += clik[is,ir,it] # already exp-ed inside updatecache_edge!
+                    ulik[is] += clik[is,ir,it] # already exp-ed inside updatecache_edge!
                 end
-                ulik[is] += u * ltw[it]
             end
         end
         end
@@ -1645,7 +1646,7 @@ function optimizegamma_LiNC!(obj::SSM, focusedge::Edge,
         updateSSM_priorltw!(obj)
         discrete_corelikelihood!(obj) # to update obj._loglikcache
     end
-    # obj._loglikcache[site,rate,tree] = log P(site | tree,rate) + log gamma(tree)
+    # obj._loglikcache[site,rate,tree] = log P(site & tree,rate)
     cadjust = maximum(view(llca, :,:,1:nt)) # to avoid underflows
     nr = length(obj.ratemodel.ratemultiplier)
     for it in 1:nt # sum likelihood over all displayed trees
@@ -1667,8 +1668,8 @@ function optimizegamma_LiNC!(obj::SSM, focusedge::Edge,
     ## step 2: Newton-Raphson
     clike ./= γ0
     clikp ./= 1.0 - γ0
-    adjustment = obj.totalsiteweight * (log(nr) - cadjust)
-    ll = obj.loglik + adjustment
+    adjustment = obj.totalsiteweight * cadjust
+    ll = obj.loglik - adjustment
     wsum = (obj.siteweight === nothing ? sum : x -> sum(obj.siteweight .* x))
     # evaluate if best γ is at the boundary: 0 or 1
     if visib # true most of the time (all the time if tree-child)
@@ -1737,7 +1738,7 @@ function optimizegamma_LiNC!(obj::SSM, focusedge::Edge,
         # if this was problematic for constraints, restrict the search
         # to interval [0,0.5] or [0.5,1] as appropriate
     end
-    ll -= adjustment
+    ll += adjustment
     obj.loglik = ll
     lγdiff = log(γ) - log(γ0); l1mγdiff = log(1.0-γ) - log(1.0-γ0)
     for it in 1:nt