… test/test_compareNetworks.jl
… test/test_phyLiNCoptimization.jl
… test/test_traitLikDiscrete.jl