new subtypes of RateVariationAcrossSites, used in fitdiscrete and in phyLiNC.