@cecileane 2020-07-01
There aren't any examples using these models in jldoctest blocks, but there are examples in the tests. For quick examples:
```julia
obj = phyLiNC!(net, fastafile, :JC69, :Inv; maxhybrid=2)
obj = phyLiNC!(net, fastafile, :JC69, :I;   maxhybrid=2) # same as above
obj = phyLiNC!(net, fastafile, :JC69, :GI;  maxhybrid=2)
obj = phyLiNC!(net, fastafile, :JC69, :GammaInv; maxhybrid=2) # same as above
```
 Eventually, we want examples in the tutorial.