@cecileane 2020-07-03
The last 3 commits are to fix the issue you found, to made similar changes to show substitution models, and to make the phyLiNC output a little prettier: see the example below. Does that sound good?

```julia
julia> obj = phyLiNC!(net_level1_s, # missing BLs, so BLs are re-estimated before starting
         fastaindiv, :JC69, :Inv; maxhybrid=2, no3cycle=true, nohybridladder=true,
         speciesfile=mappingfile, seed=106, nruns=1,
         maxmoves=10, nreject=2)
┌ Warning: Options indicate there should be no 3-cycles in the returned network,
│ but the input network contains one or more 3-cycles
│ (after removing any nodes of degree 2 including the root).
│ These 3-cycles have been removed.
└ @ PhyloNetworks ~/.julia/dev/PhyloNetworks/src/phyLiNCoptimization.jl:556
PhyLiNC network estimation starting. Parameters:
   maxhybrid = 2
   nohybridladder = true
   no3cycle = true
   probST= 0.5
   max number of moves per cycle = 10
   max number of consecutive failed proposals = 2
   optimization tolerance: ftolRel=1.0e-6, ftolAbs=1.0e-6,
                           xtolAbs=1.0e-5, xtolRel=1.0e-10.
   filename for log and err files: phyLiNC
data:
  10 species
  8 sites
  7 distinct patterns

1 run(s) starting near network topology:
((S8:0.0727550528394259,S9:0.18124625717249943):0.030293351887862727,S10:0.021238305267925087,((((S1A:0.1421303216079915,S1B:0.1272851626471764,S1C:0.27285201042241236)S1:0.37355805440121415,S4:0.18420001953480844):0.01870260718351562,(S5:0.0)#H1:0.5374708821725971::0.6):0.040639580670838275,(#H1:0.3137049331335827::0.4,(S6:0.0960770976766747,S7:0.20137114614296983):0.07654955793498787):0.058719841027722056):0.0960512809768108);
starting model:
Jukes and Cantor 69 Substitution Model,
  relative rate version
  off-diagonal rates equal to 1/3
  rate matrix Q:
                 A       C       G       T
         A       *  0.3333  0.3333  0.3333
         C  0.3333       *  0.3333  0.3333
         G  0.3333  0.3333       *  0.3333
         T  0.3333  0.3333  0.3333       *
Rate variation across sites: +I (invariable sites)
  pinv: 0.29844
  rates: [0.0, 1.425]
Time: 2020-07-03 12:42:40.019
main seed = 106
---------------------
BEGIN PhyLiNC run 1
seed = 106
time = 2020-07-03 12:42:41.606
starting at ((S8:0.0727550528394259,S9:0.18124625717249943):0.030293351887862727,S10:0.021238305267925087,((((S1A:0.1421303216079915,S1B:0.1272851626471764,S1C:0.27285201042241236)S1:0.37355805440121415,S4:0.18420001953480844):0.01870260718351562,(S5:0.0)#H1:0.5374708821725971::0.6):0.040639580670838275,(#H1:0.3137049331335827::0.4,(S6:0.0960770976766747,S7:0.20137114614296983):0.07654955793498787):0.058719841027722056):0.0960512809768108);

FINISHED. loglik = -67.68035853670689
All runs complete.
end time: 2020-07-03 12:42:46.118
time elapsed: 5.96 seconds
---------------------
Best topology:
((S8:0.32689860260599074,(S6:0.15772610798508546,S7:0.40683564720435833):0.15671654930777604):1.0e-8,((S9:0.11304098622180392,S5:0.5911111646187566):0.03834378992563829,(S4:0.21705730828551548,(S1A:1.0e-8,S1B:0.14047618838585876,S1C:0.1404838951950309)S1:1.0764437316907016):0.12275326924359169):1.0e-8,S10:1.0e-8);
with loglik -67.68035853670689 under:
Jukes and Cantor 69 Substitution Model,
  relative rate version
  off-diagonal rates equal to 1/3
  rate matrix Q:
                 A       C       G       T
         A       *  0.3333  0.3333  0.3333
         C  0.3333       *  0.3333  0.3333
         G  0.3333  0.3333       *  0.3333
         T  0.3333  0.3333  0.3333       *
Rate variation across sites: +I (invariable sites)
  pinv: 0.23753
  rates: [0.0, 1.312]
---------------------
Final optimization of branch lengths and gammas on this network... complete.
Final log-likelihood: -67.67692637551681
Final network:
((S8:0.32703020084170503,(S6:0.15711532123120855,S7:0.40713886305645625):0.1568773075454964):1.0e-8,((S9:0.10322680465032262,S5:0.5910897400283761):0.038335989222530995,(S4:0.20747999349583526,(S1A:1.0e-8,S1B:0.14044911488080697,S1C:0.14045841178305113)S1:1.0687457267281875):0.13117018839086536):1.0e-8,S10:1.0e-8);
Total time elapsed: 6.43 seconds (includes final branch length and gamma optimization)
Final time: 2020-07-03 12:42:46.587
---------------------
NOTE: This network should be interpreted as unrooted, since a likelihood model cannot identify
the root. The network can be rooted with an outgroup or other external information.
---------------------
PhyloNetworks.StatisticalSubstitutionModel:
Jukes and Cantor 69 Substitution Model,
  relative rate version
  off-diagonal rates equal to 1/3
  rate matrix Q:
                 A       C       G       T
         A       *  0.3333  0.3333  0.3333
         C  0.3333       *  0.3333  0.3333
         G  0.3333  0.3333       *  0.3333
         T  0.3333  0.3333  0.3333       *
Rate variation across sites: +I (invariable sites)
  pinv: 0.23753
  rates: [0.0, 1.312]
on a network with 0 reticulations
data:
  10 species
  8 sites
  7 distinct patterns
log-likelihood: -67.67693
```